requirements.txt
bidict
cachelib
click
Flask
Flask-Login
Flask-Session
Flask-SocketIO
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
python-engineio
python-socketio
six
Werkzeug
zipp
sphinx
pytest-cov
flask
canopen
astropy
